+++ __tests__/index.test.ts
@@ -1,0 +1,5 @@
+describe('Test', () => {
+  it('Works', () => {
+    expect(1).toBe(1);
+  });
+});