+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # CHANGE LOG
+## Unpublished
+- Update `@dylanarmstrong/nodepub` version
+- Remove unused `uuid` dependency
+- Styling on chapter title and first letter of chapter
+- Add setup code to automatically download `curl-impersonate`
+- Scaffolding for tests added
+
 ## 2023-07-04 - v0.0.8
 - Publish with types